feat(android): ti locale new property applicationLocales
