+++ tests/Resources/ti.locale.test.js
@@ -44,6 +44,12 @@ describe('Titanium.Locale', () => {
 			});
 		});
 
+		describe.android('.applicationLocales', () => {
+			it('is a Array', () => {				
+				should(Ti.Locale).have.a.property('applicationLocales').which.is.an.Array();
+			});
+		});
+
 		describe('.currentLanguage', () => {
 			it('is a String', () => {
 				should(Ti.Locale).have.a.readOnlyProperty('currentLanguage').which.is.a.String();
@@ -275,6 +281,17 @@ describe('Titanium.Locale', () => {
 
 				result = Ti.Locale.parseDecimal(undefined);
 				should(result).be.eql(Number.NaN);
+			});
+		});
+
+		describe.android('#set applicationLocales(String)', () => {
+			it('has a setter', () => {
+				should(Ti.Locale).have.a.setter('applicationLocales');
+			});
+
+			it('changes .applicationLocales', () => {
+				Ti.Locale.applicationLocales = 'ar';
+				should(Ti.Locale.applicationLocales[0].language).eql('ar');
 			});
 		});
